scm-core/src/test/java/sonia/scm/net/ahc/ByteSourceContentTest.java
/*
 * MIT License
 *
 * Copyright (c) 2020-present Cloudogu GmbH and Contributors
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */
    
package sonia.scm.net.ahc;

import com.google.common.base.Charsets;
import com.google.common.io.ByteSource;
import java.io.ByteArrayOutputStream;
import java.io.IOException;
import org.junit.Test;
import static org.junit.Assert.*;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import static org.mockito.Mockito.*;
import org.mockito.junit.MockitoJUnitRunner;


@RunWith(MockitoJUnitRunner.class)
public class ByteSourceContentTest {

  @Mock
  private AdvancedHttpRequestWithBody request;

  @Test
  public void testPrepareRequest() throws IOException
  {
    ByteSource source = ByteSource.wrap("abc".getBytes(Charsets.UTF_8));
    ByteSourceContent content = new ByteSourceContent(source);
    content.prepare(request);
    verify(request).contentLength(3l);
  }
  
  @Test
  public void testProcess() throws IOException{
    ByteSource source = ByteSource.wrap("abc".getBytes(Charsets.UTF_8));
    ByteSourceContent content = new ByteSourceContent(source);
    
    ByteArrayOutputStream baos = new ByteArrayOutputStream();
    content.process(baos);
    assertEquals("abc", baos.toString("UTF-8"));
  }

}
